+++ lib/src/flutter_taglib_bindings.dart
@@ -147,6 +147,18 @@ int taglib_bridge_has_cover(ffi.Pointer<TagLibBridgeFile> file) {
   return generated.taglib_bridge_has_cover(file);
 }
 
+int taglib_bridge_front_cover_size(ffi.Pointer<TagLibBridgeFile> file) {
+  return generated.taglib_bridge_front_cover_size(file);
+}
+
+int taglib_bridge_front_cover_data(
+  ffi.Pointer<TagLibBridgeFile> file,
+  ffi.Pointer<ffi.Uint8> buffer,
+  int bufferSize,
+) {
+  return generated.taglib_bridge_front_cover_data(file, buffer, bufferSize);
+}
+
 ffi.Pointer<TagLibBridgePictures> taglib_bridge_pictures_create() {
   return generated.taglib_bridge_pictures_create();
 }